# backend/src/main/resources/api.yaml
openapi: 3.0.0
info:
  description: |
    This is a quiz webapp server.
  version: "1.0.0"
  title: Quiz app
  contact:
    email: enver.podgorcevic@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: http://localhost:8080

tags:
  - name: quiz
    description: Quiz operations
  - name: question
    description: Question operations
  - name: answer
    description: Answer operations

paths:
  /admin/quizzes:
    get:
      tags:
        - quiz
      summary: Returns all quizzes
      description: Returns an array of all quizzes user has access to
      operationId: getAllQuizzes
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quiz'
        '401':
          description: Unauthorized quiz access
      security:
        - user_auth: []
    post:
      tags:
      - quiz
      summary: Add new quiz
      operationId: addQuiz
      responses:
        '201':
          description: Quiz created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
      security:
        - user_auth: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  description: Quiz title
                  type: string
  /admin/quizzes/{id}:
    get:
      tags:
        - quiz
      summary: Returns all the data of a quiz with matching ID
      operationId: getQuizById
      parameters:
        - name: id
          in: path
          description: Quiz ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized quiz access
        '404':
          description: Quiz not found
      security:
        - user_auth: []
    put:
      tags:
      - quiz
      summary: Update quiz by ID
      operationId: updateQuiz
      parameters:
        - name: id
          in: path
          description: Quiz ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Updated successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Resource not found
        '405':
          description: Validation exception
      security:
        - user_auth: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  description: Quiz title
                  type: string
    delete:
      tags:
        - quiz
      summary: Delete quiz by ID
      operationId: deleteQuiz
      parameters:
        - name: id
          in: path
          description: Quiz ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Resource not found
        '405':
          description: Validation exception
      security:
        - user_auth: []
  /admin/quizzes/{id}/uploadImage:
    post:
      tags:
        - quiz
      summary: Upload new quiz image
      operationId: uploadFile
      parameters:
        - name: id
          in: path
          description: Quiz ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - user_auth: []
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
  /admin/quizzes/{quizId}/swapQuestionOrder:
    put:
      tags:
        - quiz
      summary: Swap the order of two questions
      operationId: swapQuestionsOrder
      parameters:
        - name: quizId
          in: path
          description: Quiz ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Swap operation successful
        '400':
          description: Invalid question IDs
        '405':
          description: Validation exception
  /admin/quizzes/{quizId}/questions:
    post:
      tags:
        - question
      summary: Add new question
      operationId: addQuestion
      parameters:
        - name: quizId
          in: path
          description: Quiz ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: Question created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
      security:
        - user_auth: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                text:
                  description: Question text
                  type: string
                time:
                  description: Time to answer
                  type: integer
                  format: int32
                points:
                  description: Points for correct answer
                  type: integer
                  format: int32
  /admin/quizzes/{quizId}/questions/{questionId}:
    put:
      tags:
        - question
      summary: Update existing question
      operationId: updateQuestion
      parameters:
        - name: quizId
          in: path
          description: Quiz ID
          required: true
          schema:
            type: integer
            format: int64
        - name: questionId
          in: path
          description: Question ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Question not found
        '405':
          description: Validation exception      
      security:
        - user_auth: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                text:
                  description: Question text
                  type: string
                time:
                  description: Time to answer
                  type: integer
                  format: int32
                points:
                  description: Points for correct answer
                  type: integer
                  format: int32
    delete:
      tags:
        - question
      summary: Delete question by ID
      operationId: deleteQuestion
      parameters:
        - name: quizId
          in: path
          description: Quiz ID
          required: true
          schema:
            type: integer
            format: int64
        - name: questionId
          in: path
          description: Question ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Resource not found
        '405':
          description: Validation exception
      security:
        - user_auth: []
  /admin/quizzes/{quizId}/questions/{questionId}/answers:
    post:
      tags:
        - answer
      summary: Add new answer
      operationId: addAnswer
      parameters:
        - name: quizId
          in: path
          description: Quiz ID
          required: true
          schema:
            type: integer
            format: int64
        - name: questionId
          in: path
          description: Question ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: Answer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Answer'
      security:
        - user_auth: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                text:
                  description: Answer text
                  type: string
                correct:
                  description: Is answer correct
                  type: boolean
  /admin/quizzes/{quizId}/questions/{questionId}/answers/{answerId}:
    put:
      tags:
        - answer
      summary: Update existing answer
      operationId: updateAnswer
      parameters:
        - name: quizId
          in: path
          description: Quiz ID
          required: true
          schema:
            type: integer
            format: int64
        - name: questionId
          in: path
          description: Question ID
          required: true
          schema:
            type: integer
            format: int64
        - name: answerId
          in: path
          description: Answer ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Question not found
        '405':
          description: Validation exception      
      security:
        - user_auth: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                text:
                  description: Answer text
                  type: string
                correct:
                  description: Is answer correct
                  type: boolean
    delete:
      tags:
        - answer
      summary: Delete answer by ID
      operationId: deleteAnswer
      parameters:
        - name: quizId
          in: path
          description: Quiz ID
          required: true
          schema:
            type: integer
            format: int64
        - name: questionId
          in: path
          description: Question ID
          required: true
          schema:
            type: integer
            format: int64
        - name: answerId
          in: path
          description: Answer ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Resource not found
        '405':
          description: Validation exception
      security:
        - user_auth: []
externalDocs:
  description: Find out more about Swagger
  url: 'http://swagger.io'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        role:
          type: string
          description: Role of the user
          enum:
            - Admin
            - Editor
            - Player
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
    Quiz:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        picUrl:
          type: string
        creator:
          $ref: '#/components/schemas/User'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
    Question:
      type: object
      properties:
        id:
          type: integer
          format: int64
        order:
          type: integer
          format: int32
        text:
          type: string
        time:
          type: integer
          format: int32
        points:
          type: integer
          format: int32
        answers: 
          type: array
          items:
            $ref: '#/components/schemas/Answer'
    Answer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string
        correct:
          type: boolean
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
  securitySchemes:
    user_auth:
      type: http
      scheme: basic
  requestBodies:
    Quiz:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Quiz'
      description: Quiz object that needs to be created
      required: true 